added DataModel and changed interactions with the database to use DataModel.
