+++ .idea/gradle.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="GradleMigrationSettings" migrationVersion="1" />
   <component name="GradleSettings">
     <option name="linkedExternalProjectsSettings">
       <GradleProjectSettings>
@@ -8,20 +9,8 @@
         <option name="gradleJvm" value="#GRADLE_LOCAL_JAVA_HOME" />
         <option name="modules">
           <set>
-            <option value="$USER_HOME$/uni/year 2 (2024-2025)/fall 2024/CSCB07/PlanetZe_gp5" />
-            <option value="$USER_HOME$/uni/year 2 (2024-2025)/fall 2024/CSCB07/PlanetZe_gp5/app" />
-          </set>
-        </option>
-        <option name="resolveExternalAnnotations" value="false" />
-      </GradleProjectSettings>
-      <GradleProjectSettings>
-        <option name="testRunner" value="CHOOSE_PER_TEST" />
-        <option name="externalProjectPath" value="$USER_HOME$/uni/year 2 (2024-2025)/fall 2024/CSCB07/PlanetZe_gp5" />
-        <option name="gradleJvm" value="#GRADLE_LOCAL_JAVA_HOME" />
-        <option name="modules">
-          <set>
-            <option value="$USER_HOME$/uni/year 2 (2024-2025)/fall 2024/CSCB07/PlanetZe_gp5" />
-            <option value="$USER_HOME$/uni/year 2 (2024-2025)/fall 2024/CSCB07/PlanetZe_gp5/app" />
+            <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/app" />
           </set>
         </option>
         <option name="resolveExternalAnnotations" value="false" />
