+++ .idea/migrations.xml
@@ -3,7 +3,7 @@
   <component name="ProjectMigrations">
     <option name="MigrateToGradleLocalJavaHome">
       <set>
-        <option value="$USER_HOME$/uni/year 2 (2024-2025)/fall 2024/CSCB07/PlanetZe_gp5" />
+        <option value="$PROJECT_DIR$" />
       </set>
     </option>
   </component>
